Modificando servicios y repositorios para convertirlos en asincronos

+++ Birthflow_Service.Domain/DTOs/Partograph/CervicalDilationDto.cs
@@ -8,7 +8,7 @@ namespace BirthflowService.Domain.DTOs.Partograph
 {
     public class CervicalDilationDto
     {
-        public int? Id { get; set; } = null;
+        public long? Id { get; set; } = null;
         public Guid PartographId { get; set; }
         public decimal Value { get; set; }
         public DateTime Hour { get; set; }
